Update installer.sh
This script will be in a fork of its own (install) in the future instead of main and testing.

Bash indentation is 2...

Out-commented some code not necessary and made comments, this will be removed later. (mostly for the framps that created the installer in the first place)

Added a "sudo check", or the script will exit.

Made changes to renaming the config file later in script, checking if a previous file is already exists and asks for overwrite confirmation.