--rsync option
As discussed in #65 
User can now invoke `--rsync` option and get presented with the rsync line that will be used in the script to be able to modify it.

Overhaul of rsync in general. The command is now saved in the $RSYNC_COMMAND variable and ran with `eval` in the do_rsync function.